backend/requirements.txt
asgiref==3.8.1
Django==3.2.16
djangorestframework==3.12.4
django-filter==22.1
djoser==2.1.0
pillow
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1